pyproject.toml
[project]
name = "xray"
version = "0.1.0"
description = "X-Ray SDK & API - Decision-reasoning observability for multi-step pipelines"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Mohit Nagaraj", email = "mohitnagaraj20@gmail.com"}
]
dependencies = []

[project.optional-dependencies]
sdk = [
    "httpx>=0.27",
    "pydantic>=2.5",
    "pyyaml>=6.0",
]
api = [
    "fastapi>=0.109",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "aiosqlite>=0.20",
    "pyyaml>=6.0",
    "uvicorn>=0.27",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "mypy>=1.8",
    "ruff>=0.2",
]
all = ["xray[sdk,api,dev]"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared", "sdk", "api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
